refactor(channels): extract group routing + channel-layer prompts (#162)
Move the channel-agnostic pieces into core/:

- src/channels/core/routing.ts  — matchPersonaNames, decideGroupReply,
  formatGroupContext, GroupChatState (now exported), GROUP_BUFFER_MAX.
- src/channels/core/prompts.ts  — TELEGRAM_REPLY_INSTRUCTION (name kept
  verbatim), VOICE_REPLY_INSTRUCTION, CAPTURE_NUDGE_INTERVAL,
  CAPTURE_NUDGE_TEXT, captureNudgeForTurn, voiceUnavailableMessage.

channels/telegram.ts barrel re-exports all of them under their original
names; the inline engine imports them for its own use. Verbatim move — no
logic changes. Full suite: 1184 pass / 0 fail (baseline).
